Refactor Tasks to match Things app main list structure
- Transform TasksView into main list showing sections, areas, and projects
- Create TasksSectionDetailView for viewing tasks within a section/area/project
- Add Things-like navigation hierarchy:
  - Main sections (Inbox, Today, Upcoming, Anytime, Someday)
  - Logbook section (separated for completed tasks)
  - Areas with nested projects
  - Orphan projects without areas
- Extract TaskRowView into separate component for reusability
- Add task counts for all sections, areas, and projects
- Implement NavigationStack with proper value-based navigation
- Make TaskFilter Hashable for navigation support

Now matches Things app behavior: starts with organizational structure view,
tapping any item navigates to its task list
