+++ Foundation Lab/Views/Tasks/TaskRowView.swift
@@ -1,0 +1,90 @@
+//
+//  TaskRowView.swift
+//  FoundationLab
+//
+//  Created by Assistant on 7/6/25.
+//
+
+import SwiftUI
+
+struct TaskRowView: View {
+    let task: TodoTask
+    @Binding var viewModel: TasksViewModel
+    let isSelected: Bool
+    let onTap: () -> Void
+    
+    var body: some View {
+        HStack(spacing: 12) {
+            // Selection indicator
+            if viewModel.isMultiSelectMode {
+                Image(systemName: isSelected ? "checkmark.circle.fill" : "circle")
+                    .foregroundStyle(isSelected ? .blue : .secondary)
+                    .font(.title3)
+                    .contentTransition(.symbolEffect)
+            }
+            
+            // Completion button
+            Button {
+                viewModel.toggleTaskCompletion(task)
+            } label: {
+                Image(systemName: task.isCompleted ? "checkmark.circle.fill" : "circle")
+                    .foregroundStyle(task.isCompleted ? .green : .secondary)
+                    .font(.title3)
+                    .contentTransition(.symbolEffect)
+            }
+            .buttonStyle(.plain)
+            
+            // Task content
+            VStack(alignment: .leading, spacing: 4) {
+                Text(task.title)
+                    .font(.body)
+                    .strikethrough(task.isCompleted)
+                    .foregroundStyle(task.isCompleted ? .secondary : .primary)
+                
+                HStack(spacing: 8) {
+                    if task.priority != .none {
+                        Image(systemName: "flag.fill")
+                            .font(.caption2)
+                            .foregroundStyle(task.priority.color)
+                    }
+                    
+                    if let scheduledDate = task.scheduledDate {
+                        Label {
+                            Text(scheduledDate, style: .date)
+                        } icon: {
+                            Image(systemName: "calendar")
+                        }
+                        .font(.caption)
+                        .foregroundStyle(.secondary)
+                    }
+                    
+                    if !task.tags.isEmpty {
+                        HStack(spacing: 4) {
+                            ForEach(Array(task.tags), id: \.self) { tag in
+                                Text("#\(tag)")
+                                    .font(.caption)
+                                    .foregroundStyle(.blue)
+                            }
+                        }
+                    }
+                }
+            }
+            
+            Spacer()
+            
+            // Chevron for detail
+            if !viewModel.isMultiSelectMode {
+                Image(systemName: "chevron.right")
+                    .font(.caption)
+                    .foregroundStyle(.tertiary)
+            }
+        }
+        .padding(.horizontal)
+        .padding(.vertical, 12)
+        .background(isSelected ? Color.blue.opacity(0.1) : Color.clear)
+        .contentShape(Rectangle())
+        .onTapGesture {
+            onTap()
+        }
+    }
+}
